src/python/PromptHubAPI/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
tenacity==8.2.3
azure-appconfiguration-provider==1.0.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.18.2
azure.monitor.opentelemetry==1.0.0
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi